feat(teams): Add check for team update and delete

+++ server/api/teams/[id]/index.delete.ts
@@ -6,5 +6,24 @@ export default defineEventHandler(async (event) => {
 
   const {id} = await getValidatedRouterParams(event, v.parser(idParamSchema));
 
+  const team = await prisma.team.findUnique({
+    where: {id},
+    select: {
+      _count: {
+        select: {
+          members: true,
+        },
+      },
+    },
+  });
+
+  if (!team) {
+    throw createError({statusCode: 404, message: "Team introuvable."});
+  }
+
+  if (team._count.members > 0) {
+    throw createError({statusCode: 400, message: "Impossible de supprimer une équipe qui a des membres."});
+  }
+
   return prisma.team.delete({where: {id}});
 });
